Move generate_report under analysis/.
Ideally, all command lines tools should be in a separate directory, but until
that happens, the report generator tool belongs more under analysis/ than
experiment/.
